requirements.txt
python-telegram-bot==11.0.0
allennlp
mysql-connector-python